Update native bindings to node 0.12.x. (part 1)
This updates to the native bindings to work with all of Node 0.12.x's
breaking changes and cleans up some stuff, removing some macro usage
that was annoying anyway.

**This has not been tested on device yet, spi was mocked out to get
compiling to work.  Waiting on a new SPI device to do extensive testing
on device.  I will also flesh out the test suite, which is woefully
under developed**